+++ AreaShop/src/main/java/me/wiefferink/areashop/listeners/PlayerLoginLogoutListener.java
@@ -68,7 +68,6 @@ public final class PlayerLoginLogoutListener implements Listener {
 			}
 
 			// Notify admins for plugin updates
-			AreaShop.getInstance().notifyUpdate(player);
 			return false;
 		});
