--- config/checkstyle-suppressions.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0"?>
-
-<!DOCTYPE suppressions PUBLIC
-        "-//Puppy Crawl//DTD Suppressions 1.0//EN"
-        "http://www.puppycrawl.com/dtds/suppressions_1_0.dtd">
-
-<suppressions>
-    <suppress files="WorldGuardHandler6_1_3.java" checks="TypeName" />
-</suppressions>
